ml_project/requirements.txt
gdown==4.4.0
hydra-core==1.1.2
mlflow==1.25.1
protobuf~=3.19.0
numpy==1.22.3
omegaconf==2.1.2
pandas==1.4.2
PyDrive2==1.10.1
scikit-learn==1.0.2
scipy==1.8.0
pydantic==1.9.1